requirements.txt
streamlit >= 1.28.0
langchain >= 0.0.300
langchain-community >= 0.0.30
faiss-cpu >= 1.8.0
PyPDF2>=3.0.0
python-dotenv >= 1.0.0
mistralai >= 0.0.10
sentence-transformers >= 2.2.0
tiktoken >= 0.5.0
numpy >= 1.24.0
pandas >= 2.0.0